requirements.txt
aiocontextvars==0.2.2
aiofiles==0.4.0
aiohttp==3.6.2
alembic==1.4.2
appnope==0.1.0
async-generator==1.10
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
contextvars==2.4
cryptography==2.8
decorator==4.4.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.25
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.9
idna-ssl==1.1.0
immutables==0.11
importlib-metadata==1.5.2
ipython-genutils==0.2.0
jedi==0.16.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
more-itertools==8.2.0
multidict==4.7.5
Naked==0.1.31
packaging==20.3
parso==0.6.2
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
requests==2.23.0
rfc3986==1.3.2
sanic==19.12.2
sanic-openapi==0.6.0
Sanic-Plugins-Framework==0.9.2
shellescape==3.8.1
six==1.14.0
sniffio==1.1.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
Tempita==0.5.2
traitlets==4.3.3
typing-extensions==3.7.4.1
ujson==2.0.3
urllib3==1.25.8
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
yarl==1.4.2
zipp==3.1.0
